added examples
