+++ CMakeLists.txt
@@ -21,4 +21,4 @@ if(ENABLE_TESTING)
   add_subdirectory(tests)
 endif()
 
-add_subdirectory(src)
+add_subdirectory(examples)
