--- src/CMakeLists.txt
@@ -1,6 +1,0 @@
-file(GLOB_RECURSE src_list "**.cpp")
-add_executable(mraylib ${src_list})
-target_include_directories(mraylib PRIVATE ../include)
-target_include_directories(mraylib SYSTEM PRIVATE ../lib)
-target_compile_options(mraylib PRIVATE -O3)
-target_link_libraries(mraylib PRIVATE project_options project_warnings stb::stb)
